feat: Rename watchlist dialog entry and refactor form init
Export dialog as index.tsx to simplify imports across components.
Compute initialValues with useMemo and reset the react-hook-form
when initialValues change to keep the form in sync with data.
