+++ apps/front/src/features/watchlist/components/dialog/index.tsx
